+++ VIStk/Widgets/__init__.py
@@ -3,6 +3,6 @@ from VIStk.Widgets._VISMenu import Menu
 from VIStk.Widgets._QuestionWindow import QuestionWindow
 from VIStk.Widgets._MenuWindow import MenuWindow
 from VIStk.Widgets._WarningWindow import WarningWindow
-from VIStk.Widgets._Layout import Layout
 
-__all__ = ["Menu","MenuItem","MenuWindow","QuestionWindow","WarningWindow","Layout"]
+
+__all__ = ["Menu","MenuItem","MenuWindow","QuestionWindow","WarningWindow"]